Break interpreter apart into multiple delegate interpreters
The contract here is that there are four primary states for any WebSocket connection:

1. Login
2. Character creation
3. In Game
4. In a Menu

So far I've only done Login and In Game. In this version the login delegate hands off to the in game delegate after you're done logging in.

The idea going forward is that if you're using this as a MUD development kit you might want to entirely replace one of the delegates for your game to get a different command parser or a different way of authenticating users.