Move the text to the bottom
When the screen is empty and the first text appears, have it appear at the bottom next to the input box. It's where you're going to spend most of your time looking once your screen fills up so it's weird to have to look at the top for a little while at the start. It also takes your eyes far away from the input bar which could be a usability problem.